+++ src/main/kotlin/com/procyk/maciej/tcgenerator/model/ConfigurationRequest.kt
@@ -1,0 +1,8 @@
+package com.procyk.maciej.tcgenerator.model
+
+import com.procyk.maciej.tcgenerator.util.Valid
+
+interface ConfigurationRequest {
+
+    fun configure(): Valid?
+}