+++ src/main/kotlin/com/procyk/maciej/tcgenerator/model/ConfigurationService.kt
@@ -1,0 +1,6 @@
+package com.procyk.maciej.tcgenerator.model
+
+interface ConfigurationService {
+
+    fun resetState()
+}