+++ src/main/kotlin/com/procyk/maciej/tcgenerator/template/freemarker/FreemarkerTemplateParser.kt
@@ -1,15 +1,16 @@
-package com.procyk.maciej.tcgenerator.template
+package com.procyk.maciej.tcgenerator.template.freemarker
 
 import com.procyk.maciej.tcgenerator.model.TestCase
 import com.procyk.maciej.tcgenerator.model.UserInput
+import com.procyk.maciej.tcgenerator.template.TemplateParser
 import freemarker.template.Configuration
 import freemarker.template.TemplateExceptionHandler
 import java.io.File
 import java.io.StringWriter
 
-class FreemarkerTemplateParser : TemplateParser {
+class FreemarkerTemplateParser(private val templateFile: File) : TemplateParser {
 
-    override fun mergeWithTemplate(templateFile: File, testCase: TestCase, userInput: UserInput): String {
+    override fun mergeWithTemplate(testCase: TestCase, userInput: UserInput): String {
         val dataModel = mapOf(
             "testCase" to testCase,
             "userInput" to userInput.value
@@ -21,6 +22,7 @@ class FreemarkerTemplateParser : TemplateParser {
             logTemplateExceptions = false
             wrapUncheckedExceptions = true
             fallbackOnNullLoopVariable = false
+            whitespaceStripping = false
             val template = getTemplate(templateFile.name)
             val outStream = StringWriter()
             template.process(dataModel, outStream)